Break (#5)
* Starting a big effort. Better error handling based on what is in the docs, and also breaking into a lib and a main

* Removing some more functions that are already moved over

* Moving more functions to lib as methods of a TargetMachine. Need to figure out how to thread these before going much further

* I think autocomplete did that to me. Regardless, bug fixed

* Error printing to eprintln! so it correctly adheres to stdin, stdout, stderr

* Moving everything to lib and temporarily reoving threading until other things are refactored

* Compilation warnings cleaned up
